2_Backend_Football_Probability_Engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg[binary]==3.1.16
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart>=0.0.9
python-dotenv==1.0.0
numpy>=1.26.0,<2.0.0
scipy>=1.11.0,<2.0.0
pandas>=2.1.0,<3.0.0
scikit-learn>=1.4.0,<2.0.0
python-Levenshtein==0.23.0
celery==5.3.6
redis==5.0.1
requests==2.31.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
mlflow>=2.9.0

